requirements.txt
ccxt>=4.2.25
python-binance>=1.0.19
requests>=2.31.0
python-telegram-bot>=20.7
google-generativeai>=0.8.3
protobuf>=5.29.3
python-dotenv>=1.0.0
schedule>=1.2.0
selenium>=4.16.0
webdriver-manager>=4.0.1
colorlog>=6.8.0
pymysql>=1.1.0
Flask>=3.0.0
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-mock>=3.12.0
# V3 Expansion Dependencies
yfinance>=0.2.40
ta>=0.11.0
feedparser>=6.0.11
matplotlib>=3.8.0
pillow>=10.0.0

# Dashboard
Flask>=3.0.0

# Database
pymysql>=1.1.0
